requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
cx_Oracle==8.3.0
python-multipart==0.0.9

